+++ src/lib/services/evm/lcd.ts
@@ -7,6 +7,7 @@ import {
   convertAccountPubkeyToAccountAddress,
   parseWithError,
 } from "lib/utils";
+import { zEvmMsgCreate } from "../types";
 
 export const getEvmParams = (endpoint: string) =>
   axios
@@ -35,9 +36,15 @@ export const getEvmContractInfoSequencer = async (
   const tx = txs.items[0];
   const sender = convertAccountPubkeyToAccountAddress(tx.signerPubkey, prefix);
 
+  const foundMsgCreate = tx.messages.find(
+    (msg) => msg.type === "/minievm.evm.v1.MsgCreate"
+  );
+  const evmMsgCreate = zEvmMsgCreate.parse(foundMsgCreate);
+
   return {
     hash: tx.hash,
     sender,
     created: tx.created,
+    code: evmMsgCreate.code,
   };
 };